Abstract local document creation through blob storer
