+++ cmd/history2git15/main.go
@@ -3,6 +3,7 @@ package main
 import (
 	"flag"
 	"fmt"
+	"io"
 	"log"
 	"os"
 	"os/exec"
@@ -11,10 +12,24 @@ import (
 	"time"
 
 	"github.com/ddvk/rmfakecloud/internal/config"
+	"github.com/ddvk/rmfakecloud/internal/storage"
 	"github.com/ddvk/rmfakecloud/internal/storage/fs"
 	"github.com/ddvk/rmfakecloud/internal/storage/models"
 	"github.com/ddvk/rmfakecloud/internal/ui/viewmodel"
 )
+
+type BlobRemoteStorage struct {
+	directory string
+}
+
+func (brs *BlobRemoteStorage) GetRootIndex() (hash string, generation int64, err error) {
+	return brs.bs.root.GetRootIndex(brs.uid)
+}
+
+func (brs *BlobRemoteStorage) GetReader(hash string) (io.ReadCloser, error) {
+	rc, _, _, err := brs.bs.LoadBlob(brs.uid, hash)
+	return rc, err
+}
 
 func writeEntries(fd *strings.Builder, entries []viewmodel.Entry, currentLevel int) {
 	for _, e := range entries {
@@ -86,14 +101,14 @@ func main() {
 			cfg.DataDir = path.Dir(path.Dir(userdirectory))
 
 			filesystem := fs.NewStorage(cfg)
-			lbs := filesystem.BlobStorage(userdirectory)
+			lbs := storage.NewBlobStorer(filesystem, filesystem)
 
 			if tail != 0 && len(history) > tail {
 				history = history[len(history)-tail:]
 			}
 
 			for _, h := range history {
-				tree, err := h.GetHashTree(lbs)
+				tree, err := h.GetHashTree(lbs.RemoteStorage())
 				if err != nil {
 					log.Fatalf("%s: %s: %s", arg, h.Hash, err.Error())
 				}
