+++ internal/storage/fs/userstorage.go
@@ -1,24 +1,23 @@
 package fs
 
 import (
-	"bufio"
 	"errors"
 	"fmt"
 	"io"
 	"os"
-	"strings"
-	"time"
 
 	"github.com/ddvk/rmfakecloud/internal/config"
 	"github.com/ddvk/rmfakecloud/internal/model"
-	"github.com/ddvk/rmfakecloud/internal/storage"
-	"github.com/juju/fslock"
 	log "github.com/sirupsen/logrus"
 )
 
 const (
-	userDir     = "users"
-	profileName = ".userprofile"
+	cachedTreeName = ".tree"
+	profileName    = ".userprofile"
+	userDir        = "users"
+
+	// serves as root modification log and generation number source
+	historyFile = ".root.history"
 )
 
 // NewStorage new file system storage
@@ -161,87 +160,4 @@ func (fs *FileSystemStorage) RemoveUser(uid string) (err error) {
 	}
 
 	return
-}
-
-func (fs *FileSystemStorage) GetRoot(uid string) (string, int64, error) {
-	historyPath := fs.getPathFromUser(uid, historyFile)
-
-	lock := fslock.New(historyPath)
-	err := lock.LockWithTimeout(time.Duration(time.Second * 5))
-	if err != nil {
-		log.Error("cannot obtain lock")
-		return "", 0, err
-	}
-	defer lock.Unlock()
-
-	fi, err := os.Stat(historyPath)
-	if err != nil {
-		return "", 0, err
-	}
-
-	if fi.Size() == 0 {
-		return "", 0, storage.ErrorNotFound
-	}
-
-	fd, err := os.Open(historyPath)
-	if err != nil {
-		return "", 0, err
-	}
-	defer fd.Close()
-
-	scanner := bufio.NewScanner(fd)
-	lastline := ""
-	var generation int64
-	for scanner.Scan() {
-		line := scanner.Text()
-		if line != "" {
-			generation += 1
-			lastline = line
-		}
-	}
-
-	fields := strings.Fields(lastline)
-	if len(fields) != 2 {
-		return "", 0, fmt.Errorf(".root.history corrupted")
-	}
-
-	return fields[1], generation, nil
-}
-
-func (fs *FileSystemStorage) UpdateRoot(uid string, stream io.Reader, lastGen int64) (int64, error) {
-	historyPath := fs.getPathFromUser(uid, historyFile)
-
-	lock := fslock.New(historyPath)
-	err := lock.LockWithTimeout(time.Duration(time.Second * 5))
-	if err != nil {
-		log.Error("cannot obtain lock")
-		return 0, err
-	}
-	defer lock.Unlock()
-
-	currentGen := int64(0)
-	fi, err := os.Stat(historyPath)
-	if err == nil {
-		currentGen = generationFromFileSize(fi.Size())
-	}
-
-	if currentGen != lastGen && currentGen > 0 {
-		log.Warnf("wrong generation, currentGen %d, lastGen %d", currentGen, lastGen)
-		return currentGen, storage.ErrorWrongGeneration
-	}
-
-	hist, err := os.OpenFile(historyPath, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0600)
-	if err != nil {
-		return currentGen, err
-	}
-	defer hist.Close()
-
-	hist.WriteString(time.Now().UTC().Format(time.RFC3339) + " ")
-	_, err = io.Copy(hist, stream)
-	if err != nil {
-		return currentGen, err
-	}
-	hist.WriteString("\n")
-
-	return currentGen + 1, nil
 }
